lazy load default export in dev only
